backend/pyproject.toml
[project]
name = "fastapi-professional-template"
version = "0.1.0"
description = "Production-ready FastAPI backend template"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "sqlalchemy[asyncio]>=2.0.41,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "pandas>=3.0.1",
  "python-pptx>=1.0.2",
  "openpyxl>=3.1.5",
  "xlsxwriter>=3.2.9",
  "geopy>=2.4.1",
  "tqdm>=4.67.3",
  "pre-commit>=4.5.1",
  "types-requests>=2.32.4.20260107",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_return_any = true
disallow_untyped_defs = true
plugins = []

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
